--- src/main/kotlin/entry/dsm/gitauth/equusgithubauth/global/config/WebConfig.kt
@@ -1,15 +1,0 @@
-package entry.dsm.gitauth.equusgithubauth.global.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class WebConfig : WebMvcConfigurer {
-    override fun addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-            .allowedOrigins("*")
-            .allowedMethods("*")
-            .allowedHeaders("*")
-    }
-}
