+++ src/main/kotlin/entry/dsm/gitauth/equusgithubauth/global/oauth/GithubOAuth2LoginConfig.kt
@@ -24,8 +24,6 @@ class GithubOAuth2LoginConfig {
         clientRegistrationRepository: ClientRegistrationRepository,
     ) {
         http
-            .cors { it.configurationSource(corsConfigurationSource()) }
-            .csrf { it.disable() }
             .oauth2Login { oauth ->
                 oauth
                     .successHandler(githubAuthenticationSuccessHandler())
@@ -42,18 +40,5 @@ class GithubOAuth2LoginConfig {
                         authorizationEndpoint.authorizationRequestResolver(defaultResolver)
                     }
         }
-    }
-
-    @Bean
-    fun corsConfigurationSource(): CorsConfigurationSource {
-        val configuration = CorsConfiguration()
-        configuration.allowedOrigins = listOf("http://localhost:5173")
-        configuration.allowedMethods = listOf("GET", "POST", "PUT", "DELETE", "PATCH")
-        configuration.allowedHeaders = listOf("*")
-        configuration.allowCredentials = true
-
-        val source = UrlBasedCorsConfigurationSource()
-        source.registerCorsConfiguration("/**", configuration)
-        return source
     }
 }
